requirements.txt
pyyaml~=6.0.3
pymysql~=1.1.2
rich~=14.2.0
psutil~=7.1.3
gputil~=1.4.0
netifaces~=0.11.0
flask
flask_cors
flask_sqlalchemy
PyYAML
pymysql
requests